renamed strategy_wrapper.hpp to strategy_wrapper.h
