+++ applications/GeoMechanicsApplication/custom_workflows/strategy_wrapper.h
