+++ applications/GeoMechanicsApplication/tests/cpp_tests/custom_workflows/test_time_stepping.cpp
@@ -11,7 +11,7 @@
 //                   Anne van de Graaf
 //
 
-#include "custom_workflows/strategy_wrapper.hpp"
+#include "custom_workflows/strategy_wrapper.h"
 #include "custom_workflows/time_step_executor.h"
 #include "solving_strategies/strategies/solving_strategy.h"
 #include "tests/cpp_tests/geo_mechanics_fast_suite.h"
