+++ applications/GeoMechanicsApplication/tests/cpp_tests/stub_time_loop_executor.cpp
@@ -11,7 +11,7 @@
 //
 
 #include "stub_time_loop_executor.h"
-#include "custom_workflows/strategy_wrapper.hpp"
+#include "custom_workflows/strategy_wrapper.h"
 #include "custom_workflows/time_incrementor.h"
 #include "geo_mechanics_fast_suite.h"
 #include "processes/process.h"
